third_party/WebKit/LayoutTests/platform/mac-mac10.10/external/wpt/html/semantics/scripting-1/the-script-element/module/crossorigin-expected.txt

This is a testharness.js-based test.
PASS Root module, Error in CORS-same-origin script
PASS Root module, Blocked script download, missing CORS ACAO header
PASS Root module, Blocked script download, crossorigin attribute with missing CORS ACAO header
PASS Root module, Blocked script download, mismatched CORS ACAO header
PASS Imported module, Error in CORS-same-origin script
FAIL Imported module, Blocked script download, missing CORS ACAO header assert_equals: Unexpected _log value expected (string) "error" but got (undefined) undefined
PASS Imported module, Blocked script download, crossorigin attribute with missing CORS ACAO header
PASS Imported module, Blocked script download, mismatched CORS ACAO header
Harness: the test ran to completion.
